Adding a few more tests
